[FIX]Fixed JSON parsing
the res.json doesn't react well when the body is empty.
Thus we first convert to text then manually parse to JSON using
JSON.parse. Then if there is an error we can catch it and treat it
like normal and throw a warning to the user and simply return the
empty text (or another text than can't be parsed)