[DEV]Generalized Update algorithm
The onUpdate method can now extract the form data and search for
elements corresponding to a group. For instance if in the form there is
the below data
{"on-1":true, "on-2":true}
onUpdate will be able to distinguish between the 2, will do a group and
will do 2 seperate requests, merge the promises and resolve whenever
all the inner promises are resolved.
The constraint is that every attribute has to have an identifier
to be able to group. And to know if there are groups, the algorithm
looks.
Updating is also optimized to only call if the data has been altered