[FIX]Different input approach
Using a different approach as to how to understand that a form element
is a number or a text.
Whenever there is a 'type="number"' attribute, then the input is a
number. The drawback from this is that the user will have to enter
this attribute in html and if he wan't to hide the field (for example
to enter an "id" which would be a number) he will have to use the
style to set the input to "display:none" using css rules.