[DEV]Redone headers promise
Headers option now takes a promise in place of a function returning
a promise.
This is to make the errors more catchable. For instance if there is an
abort error, we can now catch it and throw away the promise.
If there is another error that the end user needs to catch, he can do
it with the catch function.