[DEV]Updated server
Installed body parser to parse requests and remove old warning in
console.
Added a custom logger to remove the parameters in console. These
parameters would take up all the space in the console and would
make you scroll all the way up if you needed to see other logs.
Added custom status code by adding the context (req,res) from
express. If the status is changed to something other than 200,
the app will never change the status code if there is an error.
Although if the status is 200 and there is an error, the app
will send a status code of 404 to the client.