[DEV]Replaced fetch & Loading & examples
Fetch has been changed to the good old XHR network handler since fetch
doesn't implement a way to get the loading status of a request/response
The QueryType loading now has a type extension which allows it to be
a boolean or also an object with the loading data attached.
You can get the download, the upload data or even the total data
progress ! You can specify at what point in percentage the uploading
is done so that the overall loading is measurable as a percentage.
Updated react because of what seemed to be a bug. It actually is a
feature :
https://github.com/facebook/react/issues/24502
Updated interfaces to match backend