[DEV]Integrated form parser & Auto header
You can now send a form event to the request event and the function
will detect that the sent data is a form event and parse it.
It will retrieve the form information that is sent and use it in the
request.
Header is determined based on the type of the variable.
If there is a form data object that is only sendable with form-data,
it will convert to a FormData. The drawback is that FormData doesn't
keep the types for numbers and boolean and simply converts them to
strings.