[DEV][FIX]Cache used for all requests
Previously, the requests were only for when using the useQuery hook.
Now for every request used inside react-query, the cache will work.
For instance, you can do a POST to X and with Y data.
X and Y will be hashed and whenever the data is retrieved again,
the cache will update.
TODO: Currently there is no way of retrieving the cache when only
using the request function... It only sets the cache.