[FIX]Optimized renderings
The render functions seamed to be called excessively. We only allow
rerenders whenever the type attribute changes. Whenever the data
changes the CRUD component (that CRUDAuto overwraps) should rerender
and it'll rerender everything down the hierarchy. Might have created
a bug for whenever the data changes. This isn't a problem for the
creation, but for the updating of the data is will be.