[DEV]Extracted code & added attributes
Added a "none" value to the query param type in case we don't want
to use an id.
Added an active attribute to Query, CRUD and CRUDAuto to setup the
loading of the query without actually using the string passed as the
endpoint. This enables to use undefined values which will be changed
when the active attribute is set to true. Set to true by default.
Extracted the UpdateForm and DeleteForm into  seperate components.
Previously used ones would call the destruction of the component
and call the component again to rebuild it. Which is not what it should
do on update.
Updated examples.